+++ cmod/protobuf/src/rmessage.c
@@ -375,7 +375,7 @@ const char* pbc_rmessage_string(pbc_rmessage* m, const char* key, int index, int
 
   if (type == PTYPE_ENUM) {
     if (sz) {
-      *sz = strlen(var->e.name);
+      *sz = (int)strlen(var->e.name);
     }
     return var->e.name;
   }
